requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cfgv==3.3.1
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
Django==3.0.8
django-cors-headers==3.8.0
django-tenant-schemas==1.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.18.1
drf-yasg==1.20.0
filelock==3.0.12
identify==2.2.13
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
nodeenv==1.6.0
ordered-set==4.0.2
packaging==21.0
platformdirs==2.3.0
pre-commit==2.15.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
